refactor(api): prune heavy runtime re-exports from core/api barrel
Slims core/api/index.ts to light entries only (constants + issues).
Heavy wrappers that construct agent classes
(runOffensiveSecurityAgent, runAttackSurfaceAgent, runPatchingAgent,
runEnvironmentAgent, runAuthenticationAgent,
runBenchmarkComparisonAgent, runPentestAgent, runTargetedPentestAgent,
runThreatModelWorkflow) are now imported from per-feature paths only.

Why: each wrapper's static import graph drags 50+ implementation
files (tools, prompts, schemas), so any import from core/api drags
every agent's full tree into the consumer's TS program. A console-
shaped probe drops from 1044 reachable files (barrel import) to 831
(deep-path import), below the pre-#725 baseline.

Internal consumers migrated to deep paths: operator-dashboard,
threatModel workflow, attack-surface/auth/blackbox-pentest scripts.
knip.json gains entries for the eight pruned wrappers.

Closes #749
