+++ knip.json
@@ -3,6 +3,15 @@
   "entry": [
     "src/cli.ts",
     "src/core/api/index.ts",
+    "src/core/api/attackSurface.ts",
+    "src/core/api/authentication.ts",
+    "src/core/api/benchmark.ts",
+    "src/core/api/blackboxPentest.ts",
+    "src/core/api/environment.ts",
+    "src/core/api/offesecAgent.ts",
+    "src/core/api/patching.ts",
+    "src/core/api/targetedPentest.ts",
+    "src/core/api/threatModel.ts",
     "scripts/**/*.ts",
     "src/core/benchmark/**/*.ts",
     "src/core/agents/specialized/benchmark/**/*.{ts,tsx}",
