+++ scripts/blackbox-pentest.ts
@@ -1,5 +1,5 @@
 import { config } from "dotenv";
-import { runPentestAgent } from "../src/core/api";
+import { runPentestAgent } from "../src/core/api/blackboxPentest";
 import { AgentEventBus } from "../src/core/eventBus";
 import { sessions } from "../src/core/session";
